refactor: auto-split test_contact.py
